Implement #13: better Animania compatibility.
* ItemEnchantedCarrot no longer works on Animania entities apart from
healing them. The message saying this is somewhat obnoxious and should
probably be set to false by default.
* ItemWhistle will now list Animania Horses:
    * Unnamed horses are listed as unsummonable.
    * Named horses with the patern "Player's Steed" will be listed as
    summonable if the player's name matches.
* ItemWhistle will now summon Animania Horses named as above.
* Sneak-right-clicking an Animania horse that is currently unnamed will
name it "Player's Steed", suiting the above conditions. Alternately, you
can use a name tag.
* If the option is set, sneak-right-clicking again with the whistle will
unname the horse, so long as it is named for you. Otherwise an error
stating the horse is already named will be printed.