Ultimate Unicorn Mod proxy.
Some other minor tweaks:
- Fix Arrays.asList -> Lists.newList and use removeIf instead.
- Move and generify onDismount event so that all EntityCreatures have
their homes cleared, not just AbstractHorse derived entities. This
should properly remove any homes set by the Ocarina.

Unicorn Proxy:

Specifically handle Entity Deaths (test this actually
functions properly without Unicorn installed) of the Vagabond to set
ownership of its mount to the killing player (if there is one).

This is in addition to the changes to the Enchanted Carrot that allows
you to tame animals that are already tamed but don't currently have an
owner.

Considering all of those are releated to the Unicorn mod, I've also
included TextComponents for the list functionality of the Ocarina to
note if there are nearby tamed, un-owned horses (and is there is one
being ridden by a Vagabond).

Properly handle the unique names of the unique Unicorns so that it just
displays the base class name and not "(named Alea)" but also support
actually renaming them.

Resolves #24.