Re-implement sound play points for Ocarina; tweak wording.
Now, a "broken" whistle sound will play when the ocarina is at its min
durability (and that durability is not 0).

"Minor" key whistles will play when a summoning fails part of the way
through.

Otherwise, a major key whistle is played while additionally still having
the chance of being the Epona easter egg.