Resolve #40: Mo Creatures names should now be demangled properly.
This restores the code that adds additional spaces to the entities
listed in the configuration.

It specifically only firest on TextComponentTranslations with the
specific key of "dwmh.strings.is_at", and modifies the format arguments
for that before piecing together a new component.

Utterly not the best way to do it and will be resolved in a future
version more properly.