Specify that the glint is currently a client-side only option.
i.e., the setting on the server makes no difference.

Future plan is to split into server/client configuration so that it's
easier to tell what settings people can change on their own when playing
on multiplayer.

However, that means network packets for entities, messages, sound
events, etc, are all transferred instead of ITextComponent (or in
addition to) messages being sent.