Ensure error messages are logged for invalid configuration items specifically. (#51)
Store ItemStack copies of the defaults as well.

To ensure that there's no registration issues (as these would otherwise
be called on init with something like, updateConfiguration), I added a
new function which merely calls parseItem with a new boolean flag
indicating that it should report errors.

Otherwise errors are reported constantly with an invalid configuration
when looking at the tooltip of the carrot/ocarina for example.