Adjacent to #58, resolve AbstractHorse-derived entities with disabled proxies.
Animania horses and Ultimate Unicorn Mod horses all derive from
AbstractHorse, meaning that even with their proxies disabled, it doesn't
actually "disable" them.

This modifies the vanilla isMyMod to add them to the ignore list. Which
is hopefully quicker. I don't know. It's not called often enough to be
an issue.