Conditionally register recipes for saddle & Enchanted Carrot.
Resolves #57 and resolve #56. The Enchanted Carrot is now always
registered, but the recipe itself can be disabled through the
configuration.

Additionally, setting "enabled" to "false" for the Enchanted Carrot no
longer has an affect on any of its functionality, meaning it can still
be used as an OP tool if cheated in, or it can be used in loot bags &c
without allowing crafting.

The saddle recipe is the same as I've used in my work-in-progress
modpack. I feel as though horses should be accessible much earlier than
simply getting lucky with loot in a chest or the right villager.

Considering how heavily saddles are themed around the mod, adding a
recipe for crafting them makes sense. The recipe is well-balanced: a
"hardened" leather would be preferred but I have no desire to make
additional materials (or textures), and the combination of the
ingot->block->ingot hopefully indicates (at least visually, and at least
one) stirrups.