Resolve #52: check for item breakage while summoning.
In combination with the changes to make it work with Unbreaking, this
also means that you can't summon a whole horde of eligible entities with
only 1 durability, break the item after the first entity, and then
continue summoning.

Instead, it will now stop mid-summon and present the player with a
message.