Revert item damage changes, resolve #63: breakable carrot.
Turns out damageItem already calls the relevant function, I'm not sure
what made me think that it didn't. Which meant that it never actually
broke the item. Oops!

Resolving that and adding an option (false by default, currently) which
makes enchanted carrots a consumable again. Their icon correctly changes
and they correctly break when their durability is used up, which is a
win.